requirements.txt
mysql-connector-python
dotenv
openai
sqlparse
typing